The primary differences between the last version and the actual version are:
In the actual version, the private user: User | null; is added to store the actual user.
The constructor of the actual version adds a new dependency injection for private alertService: AlertService.
In the ngOnInit() method, the actual version replaces let actualUSer: User | null with this.user.
In the getPageResponse() method, the actual version adds a new method this.setModelKeys(keys); to set the model keys, and it replaces this.modelKeys with let keys: string[].
The actual version adds a new private method private setModelKeys(responseKeys: string[]): void to set the model keys based on the response keys and user roles.
The actual version removes the console.log(this.pageRequest) statement from the getPageResponse() method.
The actual version initializes the modelKeys array in its declaration as an empty array public modelKeys: string[] = [];.
The actual version introduces a more organized approach to setting the model keys by adding the setModelKeys method, handling user roles, and managing hidden keys based on the user's role. Additionally, the actual version includes an AlertService that provides a better user experience by displaying relevant alerts or messages.