feat: Add breadcrumb and add new category feature
This commit adds breadcrumb functionality to the category page, allowing the user to navigate back to the main page or any previous category page. Additionally, a new feature has been added that allows the user to add a new category to the list. The design of the category page has also been updated to make the category cards clickable, which will take the user to the corresponding category page.